jetson/requirements.txt
numpy==1.19.5
pandas==1.3.3
matplotlib==3.4.3
scikit-learn==0.24.2
opencv-python>=4.1
mtcnn==0.1.0
imutils==0.5.4
wand==0.6.7
Pillow==8.4.0
# tensorflow==2.6.0